Small refactoring in flaskr: move logic out of __init__.py into separate file
